delete some redundant scripts and codes
